dev-requirements.txt
-r requirements.txt

mock==1.0.1
pytest==2.7.3
httpretty==0.8.4
pytest-cov==1.8.1
pytest-django==2.8.0
ipdb==0.8
ipython==3.1.0

pep8>=1.5.7,<1.6.0
pyflakes>=0.8,<0.9
mccabe>=0.3,<0.4
flake8==2.3.0
freezegun==0.3.2
